Server/requirements.txt
Flask
pymongo
ollama
python-dotenv
flask-swagger-ui
pyyaml
